+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeTemplateTag.assignment` + `EnvelopeTemplateTag.EnvelopeRecipient` can be null
 - Add `EnvelopeTag.assignment` + `EnvelopeTag.EnvelopeRecipient` can be null
 - Add `EnvelopeTemplate.useDefaultTemplateSettings`
-- Add `EnvelopeDocument.stamperPosition`, `EnvelopeDocument.stamperPositionX`, `EnvelopeDocument.stamperPositionY`, 
+- Add `EnvelopeDocument.labelPosition`, `EnvelopeDocument.labelPositionX`, `EnvelopeDocument.labelPositionY`, 
 
 ## [1.4.0] - 2021-09-17
 ### Added